Time: 21 ms (7.23%), Space: 49.6 MB (48.32%) - LeetHub

+++ 2090-number-of-ways-to-arrive-at-destination/2090-number-of-ways-to-arrive-at-destination.java
@@ -44,7 +44,7 @@ class Solution {
       }
      }
 
-     return (int)c[n-1];
+     return c[n-1];
